fix: Move the macros-gate pin into src/tests, outside coverage's reach
The codecov/patch failures on this branch were phantom: every one of
the flagged lines was inside the new pinning test, whose in-module
test lines the coverage post-processing strips on Linux but not on
Windows (its module detection finds 85 test modules there against 81
on Windows), so the merged report counted test lines only some uploads
carried — and the count of "missing" lines grew with each test line
added. CI already excludes parser/src/tests/** from the lcov at
generation on every platform (--ignore-filename-regex tests), so the
pin now lives there, reaching the gate through a test-only pub(crate)
re-export, and the patch is production lines alone.
