+++ parser/src/tests/mod.rs
@@ -10,6 +10,7 @@ mod block_nesting_depth;
 pub(crate) mod fixtures;
 mod hash;
 mod inline_builder_document_parity;
+mod inline_builder_macro_gate;
 mod inline_builder_passthrough_record_parity;
 mod inline_builder_side_effect_parity;
 mod inline_renderer;
